pyproject.toml
[project]
name = "ufpi-documents-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Luis Enilton",email = "luis.enilton@visagio.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "langchain-google-genai (>=2.1.10,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "pypdf (>=6.0.0,<7.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "pymupdf (>=1.26.4,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "unstructured[ocr,pdf] (>=0.18.14,<0.19.0)",
    "sentence-transformers (>=2.7.0,<3.0.0)",
    "langchain-openai (>=0.2.12,<0.3.0)",
    "langchain-huggingface (>=0.1.0,<0.2.0)",
    "torch (>=2.0.0)",
    "transformers (>=4.38.0)"
]
[tool.poetry]
package-mode = false
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
